requirements.txt
ccxt==4.4.75
python-telegram-bot==21.4
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0